% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{munnell}
\alias{munnell}
\alias{munnell_wide}
\title{Productivity in the United States}
\format{
a data frame containing:
\itemize{
\item state: the state
\item year: the year
\item region: one of the 9 regions of the United States
\item president: the name of the president for the given year
\item publiccap: public capital stock
\item highway: highway and streets
\item water: water and sewer facilities
\item utilities: othe public building and structures
\item privatecap: private capital stock
\item gsp: gross state product
\item labor: labor input measured by the employment in non--agricultural payrolls
\item unemp: state unemployment rate
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 48 rows and 36 columns.
}
\source{
Online complements to Baltagi (2001): \url{https://www.wiley.com/legacy/wileychi/baltagi/}
Online complements to Baltagi (2013): \url{https://bcs.wiley.com/he-bcs/Books?action=resource&bcsId=4338&itemId=1118672321&resourceId=13452}
}
\usage{
munnell

munnell_wide
}
\description{
a panel data of 48 American States for 17 years, from 1970 to 1986
}
\references{
\insertRef{BALT:01}{dfidx}

\insertRef{BALT:13}{dfidx}

\insertRef{BALT:PINN:95}{dfidx}

\insertRef{MUNN:90}{dfidx}
}
\keyword{dataset}
\keyword{datasets}
