% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfidx.R
\name{unfold_idx}
\alias{unfold_idx}
\alias{fold_idx}
\title{Fold and Unfold a dfidx object}
\usage{
unfold_idx(x)

fold_idx(x, pkg = NULL)
}
\arguments{
\item{x}{a \code{dfidx} object}

\item{pkg}{a pakcage}
}
\description{
\code{fold_idx} takes a dfidx, includes the index as stand alone
columns, remove the \code{idx} column and return a data.frame, with an
\code{ids} attribute. \code{fold_idx} performs the opposite operation
}
\examples{
data("TravelMode", package = "AER")
TM <- dfidx(TravelMode)
TM2 <- unfold_idx(TM)
attr(TM2, "ids")
TM3 <- fold_idx(TM2)
identical(TM, TM3)
}
\author{
Yves Croissant
}
