% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as_tibble}
\alias{\%>\%}
\alias{filter}
\alias{arrange}
\alias{slice}
\alias{pull}
\alias{mutate}
\alias{transmute}
\alias{select}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr:reexports]{\%>\%}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr:reexports]{as_tibble}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{pull}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{slice}}, \code{\link[dplyr]{transmute}}}
}}

